Require superuser permission to revoke global roles
Correctly handle deleting team global roles and shared global roles

Comment adjusting

Disallow view exception for system roles
