+++ ansible_base/rbac/api/serializers.py
@@ -145,7 +145,7 @@ class RoleDefinitionSerializer(CommonModelSerializer):
         else:
             content_type = self.instance.content_type
         validate_permissions_for_model(permissions, content_type)
-        check_locally_managed(permissions)
+        check_locally_managed(permissions, content_type)
         return super().validate(validated_data)
 
 
@@ -250,13 +250,15 @@ class BaseAssignmentSerializer(CommonModelSerializer):
         if getattr(obj, 'validate_role_assignment', None):
             obj.validate_role_assignment(actor, rd)
 
+        # Return a 400 if the role is not managed locally
+        check_locally_managed(rd.permissions.prefetch_related('content_type'), rd.content_type)
+
         if rd.content_type:
             # Object role assignment
             if not obj:
                 raise ValidationError({'object_id': _('Object must be specified for this role assignment')})
 
             check_content_obj_permission(requesting_user, obj)
-            check_locally_managed(rd.permissions.prefetch_related('content_type'))
 
             try:
                 with transaction.atomic():
